+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - App icon on Windows
 - Upgraded dependencies
 
+### Added:
+- Category color text field for HEX color, with ColorPicker moved to bottom sheet
+
 ## [0.8.319] - 2023-04-13
 ### Changed:
 - Measurable setting page layout